:lipstick: Update bit manipulation style
Further explains how bit modify should be used in the code base, such as
preferring to use the APIs with bit mask template arguments to ensure
the call knows the bit mask at compile time.

Also provides a carve out for concatenating bits together. In the case
where the developer simply needs to combine a set of bits together with
an OR operation, these are simple enough to read and understand and do
not require the bit APIs to be used.

Resolves #53